requirements.txt
openai>=1.14.0
google-api-python-client>=2.118.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
moviepy>=1.0.3
Pillow>=10.2.0
pytrends>=4.9.2
pydub>=0.25.1
requests>=2.31.0
mutagen>=1.47.0
